# Optimize getCooccurrence for better performance