Experimental

```
Input: 主角長相畫風被越畫越像路人？
=== New ===
最為相似的標題是 ['主角', '長', '相', '畫', '風', '被', '越', '畫', '越', '像', '路人'] ，相似度為 388 
Take: 11490440453 ns
=== Old ===
最為相似的標題是 ['主角', '長', '相', '畫', '風', '被', '越', '畫', '越', '像', '路人'] ，相似度為 489 
Take: 11986769524 ns

---

Input: 超商的新進員工實習期間有幾天
=== New ===
最為相似的標題是 ['超', '商', '的', '新', '進', '員', '工', '實', '習', '期', '間', '有', '幾', '天'] ，相似度為 629 
Take: 12585262843 ns

=== Old ===
最為相似的標題是 ['超', '商', '的', '新', '進', '員', '工', '實', '習', '期', '間', '有', '幾', '天'] ，相似度為 629 
Take: 14101854292 ns
```